LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/header-parsing.https-expected.txt

CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "https://localhost:9443" from accessing a frame at "https://localhost:9443".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "https://localhost:9443" from accessing a frame at "https://localhost:9443".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "https://localhost:9443" from accessing a frame at "https://localhost:9443".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "https://localhost:9443" from accessing a frame at "https://localhost:9443".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "https://localhost:9443" from accessing a frame at "https://localhost:9443".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "https://localhost:9443" from accessing a frame at "https://localhost:9443".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "https://localhost:9443" from accessing a frame at "https://localhost:9443".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "https://localhost:9443" from accessing a frame at "https://localhost:9443".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.

PASS navigation allowed for []
PASS navigation allowed for [""]
PASS navigation allowed for ["jibberish"]
PASS navigation allowed for [{"percentEncoded":"require%FFcorp"}]
PASS navigation allowed for ["require-corp;"]
PASS navigation allowed for ["\u000brequire-corp\u000b"]
PASS navigation allowed for ["\frequire-corp\f"]
PASS navigation allowed for ["\rrequire-corp\r"]
PASS navigation allowed for ["Require-corp"]
PASS navigation allowed for ["\"require-corp\""]
PASS navigation allowed for [":cmVxdWlyZS1jb3Jw:"]
PASS navigation allowed for ["require-corp;\tfoo=bar"]
PASS navigation allowed for ["require-corp require-corp"]
PASS navigation allowed for ["require-corp,require-corp"]
PASS navigation allowed for ["require-corp","require-corp"]
PASS navigation allowed for ["","require-corp"]
PASS navigation allowed for ["require-corp",""]
PASS navigation blocked for ["require-corp"]
PASS navigation blocked for [" require-corp "]
PASS navigation blocked for ["\trequire-corp\t"]
PASS navigation blocked for [" \trequire-corp"]
PASS navigation blocked for ["require-corp\t "]
PASS navigation blocked for ["require-corp; foo=bar"]
PASS navigation blocked for ["require-corp;require-corp"]
PASS navigation blocked for ["require-corp; report-to=\"data:","\""]
